move core functions to core package
